Selection Background (#19)
When hovering a class or function name, the ide will show you some info about it and change the background color of the whole block to `@selection-color`. As it is, sky blue is not very suitable. Here is a much more subtle brown. I think it fits the rest of the theme much better.